+++ src/components/Message/Message.js
@@ -1,0 +1,14 @@
+
+import './Message.css'
+
+function Message({message}) {
+  return (
+    <p className="message">
+    {message}
+    </p>
+  );
+}
+
+
+export default Message;
+
