+++ src/components/Timestamp/Timestamp.js
@@ -1,0 +1,13 @@
+
+import './Timestamp.css'
+
+function Timestamp({timestamp}) {
+  return (
+    <span className="timestamp">{timestamp}
+    </span>
+  );
+}
+
+
+export default Timestamp;
+
