+++ src/components/User/User.css
@@ -1,0 +1,87 @@
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
+    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+.App {
+  padding: 20px;
+}
+
+.tweet {
+  margin: 0 auto;
+  min-width: 480px;
+  max-width: 640px;
+  min-height: 80px;
+  padding: 10px;
+  border: 1px solid #ccc;
+  display: flex;
+  font-size: 1.1rem;
+  line-height: 18px;
+}
+
+.body {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
+
+.top {
+  display: flex;
+  line-height: 1.4rem;
+}
+
+.more {
+  margin-left: 20px;
+}
+
+.profile {
+  width: 50px;
+  height: 50px;
+  border-radius: 50px;
+  margin-right: 10px;
+}
+
+.name {
+  font-weight: bold;
+  margin-bottom: 0.5em;
+  margin-right: 0.3em;
+}
+
+.handle {
+  color: #536471;
+  margin-left: 3px;
+  margin-right: 3px;
+}
+
+.timestamp {
+  color: #536471;
+  font-size: 1.1rem;
+}
+
+.timestamp::before {
+  content: "·";
+  padding-right: 0.3em;
+  font-size: 1.1rem;
+  font-weight: 600;
+}
+
+.message {
+  text-align: left;
+  font-stretch: 100%;
+  font-weight: 400;
+  line-height: 1.6rem;
+  overflow-wrap: break-word;
+  white-space: pre-wrap;
+}
+
+.actions {
+  padding: 5px 5px;
+  width: 360px;
+  display: flex;
+  justify-content: space-between;
+  color: #536471;
+}
