+++ src/components/User/User.js
@@ -1,0 +1,17 @@
+
+import './User.css'
+
+function User({userData}) {
+  const {name, handled} = userData
+
+  return (
+    <span className="user">
+    <span className="name">{name}</span>
+    <span className="handle">{handled}</span>
+  </span>
+  );
+}
+
+
+export default User;
+
